requirements.txt
web3==6.0.0
python-dotenv==1.0.0
aave-v3-py==0.5.0
chainlink-contracts==0.1.0
openai==1.3.0
pandas==2.0.0
numpy==1.24.0
requests==2.28.2
pytest==7.3.1
